safari-534.53-branch/LayoutTests/fast/js/kde/object_prototype-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sub.x is 1
PASS sub.y is 2
PASS sub.hasOwnProperty('x') is false
PASS sub.hasOwnProperty('y') is true
PASS sub.x is 6
PASS sub.hasOwnProperty('x') is true
PASS sub.y is undefined
PASS sub.hasOwnProperty('y') is false
PASS obj.hasClass1 is true
PASS obj.hasClass2 is true
PASS obj.hasClass3 is true
PASS Class1.prototype.isPrototypeOf(obj) is true
PASS Class2.prototype.isPrototypeOf(obj) is true
PASS Class3.prototype.isPrototypeOf(obj) is true
PASS obj.isPrototypeOf(Class1.prototype) is false
PASS obj.isPrototypeOf(Class2.prototype) is false
PASS obj.isPrototypeOf(Class3.prototype) is false
PASS Class1.prototype.isPrototypeOf(Class2.prototype) is true
PASS Class2.prototype.isPrototypeOf(Class1.prototype) is false
PASS Class1.prototype.isPrototypeOf(Class3.prototype) is true
PASS Class3.prototype.isPrototypeOf(Class1.prototype) is false
PASS Class2.prototype.isPrototypeOf(Class3.prototype) is true
PASS Class3.prototype.isPrototypeOf(Class2.prototype) is false
PASS Class1.prototype.prototype is undefined.
PASS myfunc.length is 3
PASS myfunc.someproperty is 4
PASS myfunc.propertyIsEnumerable('length') is false
PASS myfunc.propertyIsEnumerable('someproperty') is true
PASS checkEnumerable(myfunc,'length') is false
PASS checkEnumerable(myfunc,'someproperty') is true
PASS successfullyParsed is true

TEST COMPLETE
